rebase and clean to properly merge
