--- mainApp/templates/about.html
@@ -1,16 +1,0 @@
-<!-- handles about page content -->
-{% extends "layout.html" %}
-
-{% block content %}
-
-<!DOCTYPE html>
-<html>
-<body>
-
-<h1>temporary about page</h1>
-<p>temporary about page</p>
-
-</body>
-</html>
-
-{% endblock content %}
