--- mainApp/templates/buy.html
@@ -1,16 +1,0 @@
-<!-- handles buy page content -->
-{% extends "layout.html" %}
-
-{% block content %}
-
-<!DOCTYPE html>
-<html>
-<body>
-
-<h1>temporary buy page</h1>
-<p>temporary buy page</p>
-
-</body>
-</html>
-
-{% endblock content %}
