--- mainApp/templates/layout.html
@@ -1,88 +1,0 @@
-<!-- Layout templete of common page traits thats extended by other pages to prevent code being repeated  -->
-
-<!DOCTYPE html>
-<html>
-
-<head>
-    <!-- bootstrap requirements -->
-    <meta charset="utf-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
-    <!-- favicon (icon in tab)  -->
-    <link rel="icon" type="image/png" href="{{ url_for('static', filename='favicon-16x16.png') }}">
-    <!-- bootstrap requirements -->
-    <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/css/bootstrap.min.css"
-        integrity="sha384-ggOyR0iXCbMQv3Xipma34MD+dH/1fQ784/j6cY/iJTQUOhcWr7x9JvoRxT2MZw1T" crossorigin="anonymous">
-    <link rel="stylesheet" type="text/css" href="{{ url_for('static', filename='main.css') }}">
-    {% block css %}
-
-    {% endblock %}
-    <!-- change page title as needed -->
-    {% if title %}
-    <title>Spartan Textbooks - {{ title }}</title>
-    {% else %}
-    <title>Spartan Textbooks</title>
-    {% endif %}
-</head>
-
-<!-- Navigation Bar -->
-
-<body>
-    <header class="site-header">
-        <nav class="navbar navbar-expand-md navbar-dark bg-dark fixed-top">
-            <div class="container">
-                <a class="navbar-brand mr-4" href="/">Spartan Books</a>
-                <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarToggle"
-                    aria-controls="navbarToggle" aria-expanded="false" aria-label="Toggle navigation">
-                    <span class="navbar-toggler-icon"></span>
-                </button>
-                <div class="collapse navbar-collapse" id="navbarToggle">
-                    <div class="navbar-nav mr-auto">
-                        <a class="nav-item nav-link" href="/">Home</a>
-                        <a class="nav-item nav-link" href="/about">About</a>
-                        <a class="nav-item nav-link" href="/sell">Sell</a>
-                        <a class="nav-item nav-link" href="/buy">Buy</a>
-                    </div>
-                    <!-- change what shows up in navbar based on if user is logged in or not -->
-                    <div class="navbar-nav">
-                        {% if current_user.is_authenticated %}
-                        <a class="nav-item nav-link" href="{{ url_for('logout_page') }}">Logout</a>
-                        <a class="nav-item nav-link" href="{{ url_for('profile_page') }}">Profile</a>
-                        {% else %}
-                        <a class="btn btn-pill btn-primary" href="{{ url_for('login_page') }}">Login</a>
-                        <a class="nav-item nav-link" href="{{ url_for('register_page') }}">Register</a>
-                        {% endif %}
-                    </div>
-                </div>
-            </div>
-        </nav>
-    </header>
-    <!-- warning message popup -->
-    <main role="main" class="container">
-        <div class="row">
-            <div class="col-md-12 mt-5">
-                {% with messages =get_flashed_messages(with_categories=true) %}
-                {%if messages%}
-                {%for category, message in messages%}
-                <div class="alert alert-{{category}}">
-                    {{message}}
-                </div>
-                {% endfor %}
-                {% endif %}
-                {%endwith%}
-                {% block content %}{% endblock %}
-            </div>
-        </div>
-    </main>
-    <!-- bootstrap requirements -->
-    <script src="https://code.jquery.com/jquery-3.3.1.slim.min.js"
-        integrity="sha384-q8i/X+965DzO0rT7abK41JStQIAqVgRVzpbzo5smXKp4YfRvH+8abtTE1Pi6jizo"
-        crossorigin="anonymous"></script>
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.14.7/umd/popper.min.js"
-        integrity="sha384-UO2eT0CpHqdSJQ6hJty5KVphtPhzWj9WO1clHTMGa3JDZwrnQq4sF86dIHNDz0W1"
-        crossorigin="anonymous"></script>
-    <script src="https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/js/bootstrap.min.js"
-        integrity="sha384-JjSmVgyd0p3pXB1rRibZUAYoIIy6OrQ6VrjIEaFf/nJGzIxFDsf4x0xIM+B07jRM"
-        crossorigin="anonymous"></script>
-</body>
-
-</html>
