--- mainApp/templates/profile.html
@@ -1,59 +1,0 @@
-<!-- handles user account page content -->
-{% extends "layout.html" %}
-
-{% block css %}
-<link rel="stylesheet" type="text/css" href="{{ url_for('static', filename='profile.css') }}">
-{% endblock %}
-
-{% block content %}
-    <div class="container mt-5">
-    <div class="main-body">
-
-        <div class="row gutters-sm">
-            <div class="col-md-3 mb-3">
-                <div class="card">
-                    <div class="card-body">
-                        <div class="d-flex flex-column align-items-center text-center">
-                            <img src="{{ url_for('static', filename='uploads/'+ profile.image_file) }}" alt="Admin"
-                                 class="rounded-circle" width="150">
-                            <div class="mt-3">
-                                <h4>{{ profile.username }}</h4>
-                                <p class="text-secondary mb-1">Email for Contact</p>
-                                <p class="text-muted font-size-sm">{{ profile.email }}</p>
-                                {#                                    <button class="btn btn-primary">Follow</button>#}
-                                {#                                    <button class="btn btn-outline-primary">Message</button>#}
-                            </div>
-                        </div>
-                    </div>
-                </div>
-            </div>
-            <div class="col-md-9">
-                <div class="card mb-3">
-                    <div class="card-group">
-                        <div class="row">
-                            {% for post in posts %}
-                                <div class="col-md-3">
-                                    <div class="card" style="width: 11rem; ">
-                                        <img class="card-img-top"
-                                             src="{{ url_for('static', filename='posts/'+post.cover_photo) }}"
-                                             class="img-fluid"
-                                             alt="book cover">
-                                        <ul class="list-group list-group-flush">
-                                            <li class="list-group-item">Book Title: {{ post.title }}</li>
-                                            <li class="list-group-item">ISBN-10: {{ post.isbn }}</li>
-                                            <li class="list-group-item">Condition: {{ post.book_condition }}</li>
-                                        </ul>
-                                        <div class="card-body">
-                                            <a href="delete_post/{{ post.id }}" class="card-link">Delete</a>
-                                        </div>
-                                    </div>
-                                </div>
-                            {% endfor %}
-                        </div>
-                    </div>
-                </div>
-            </div>
-        </div>
-    </div>
-{% endblock %}
-
