--- runApp.py
@@ -1,4 +1,0 @@
-from mainApp import app
-# run the app in debug mode, which allows us to see any errors of flask in the browser
-if __name__ == '__main__':
-    app.run(debug=True)
