+++ src/mainApp/__init__.py
@@ -4,7 +4,9 @@ from flask_bcrypt import Bcrypt
 from flask_login import LoginManager
 
 app = Flask(__name__)
+UPLOAD_FOLDER = './uploads/'
 app.config['SECRET_KEY'] = 'ma3b33ipq3pqr21739cf9a8'
+app.config['UPLOAD_FOLDER'] = UPLOAD_FOLDER
 app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:///site.db'
 db = SQLAlchemy(app)
 bcrypt = Bcrypt(app)
