+++ src/mainApp/models.py
@@ -27,13 +27,15 @@ class Post(db.Model):
     """User post model"""
     id = db.Column(db.Integer, primary_key=True)
     title = db.Column(db.String(100), nullable=False)
-    isbn = db.Column(db.String(10), nullable=False)
+    isbn = db.Column(db.String(100), nullable=False)
     book_price = db.Column(db.String(4), nullable=False)
     book_condition = db.Column(db.String(10), nullable=False)
     date_posted = db.Column(db.DateTime, nullable=False,
                             default=datetime.utcnow)
-    content = db.Column(db.Text, nullable=False)
+    contact = db.Column(db.String(100), nullable=False)
+    cover_photo = db.Column(db.String(20), nullable=False,
+                           default='default.jpg')
     user_id = db.Column(db.Integer, db.ForeignKey('user.id'), nullable=False)
 
     def __repr__(self):
-        return f"Post('{self.title}', '{self.date_posted}')"
+        return f"Post('{self.title}', '{self.isbn}', '{self.date_posted}', '{self.book_price}', '{self.book_condition}', '{self.contact}', '{self.cover_photo}', '{self.user_id}')"
