+++ src/mainApp/routes.py
@@ -1,15 +1,25 @@
 # main flask app
 from flask import render_template, url_for, flash, redirect, request
 from mainApp import app, db, bcrypt
-from mainApp.forms import RegistrationForm, LoginForm
+from mainApp.forms import RegistrationForm, LoginForm, SellForm
 from mainApp.models import User, Post
 from flask_login import login_user, current_user, logout_user, login_required
+from werkzeug.utils import secure_filename
+import os
+
+basedir = os.path.abspath(os.path.dirname(__file__))
+ALLOWED_EXTENSIONS = {'jpg', 'jpeg'}
+
+
+def allowed_file(filename):
+    return '.' in filename and filename.rsplit('.', 1)[1].lower() in ALLOWED_EXTENSIONS
 
 
 @app.route('/')
 @app.route('/home')
 def home_page():
-    return render_template('home.html')
+    posts = Post.query.all()
+    return render_template('home.html', posts=posts)
 
 
 @app.route('/about')
@@ -45,9 +55,21 @@ def register_page():
         # take user input password and hashit
         hashed_password = bcrypt.generate_password_hash(
             form.password.data).decode('utf-8')
+        f = form.profile_picture.data
+        if allowed_file(f.filename):
+            filename = secure_filename(f.filename)
+            if os.path.exists('./static/uploads/' + filename) != True:
+                f.save(os.path.join(basedir, 'static/uploads', filename))
+            else:
+                flash('File with same name already Existss')
+                return render_template('register.html', title='Register', form=form)
+        else:
+            flash('This extension is not allowed')
+            return render_template('register.html', title='Register', form=form)
+
         # create a database entry to register new user
         user = User(username=form.username.data,
-                    email=form.email.data, password=hashed_password)
+                    email=form.email.data, password=hashed_password, image_file=filename)
         # add user to database
         db.session.add(user)
         db.session.commit()
@@ -61,18 +83,52 @@ def buy_page():
     return render_template('buy.html', title='Buy')
 
 
-@app.route('/sell')
+@app.route('/sell', methods=['GET', 'POST'])
+@login_required
 def sell_page():
-    return render_template('sell.html', title='Sell')
+    form = SellForm()
+    if form.validate_on_submit():
+        f = form.cover_page.data
+        if allowed_file(f.filename):
+            filename = secure_filename(f.filename)
+            if os.path.exists('./static/posts/' + filename) != True:
+                f.save(os.path.join(basedir, 'static/posts', filename))
+                post = Post(title=form.title.data, isbn=form.isbn.data, book_condition=form.book_condition.data,
+                            book_price=form.book_price.data, contact=form.contact.data, cover_photo=filename,
+                            user_id=current_user.get_id())
+                db.session.add(post)
+                db.session.commit()
+                flash('Post Added Successfully', 'success')
+                return redirect(url_for('home_page'))
+            else:
+                flash('File with same name already Exists')
+                return render_template('sell.html', title='Sell', form=form)
+        else:
+            flash('This extension is not allowed')
+            return render_template('sell.html', title='Sell', form=form)
+    return render_template('sell.html', title='Sell', form=form)
 
 
 @app.route("/logout")
+@login_required
 def logout_page():
     logout_user()
     return redirect(url_for('home_page'))
+
+@app.route("/delete_post/<id>", methods=['GET', 'POST'])
+@login_required
+def delete_post(id):
+    Post.query.filter_by(id=id).delete()
+    db.session.commit()
+    flash('Post Deleted Successfully')
+    return redirect(url_for('profile_page'))    
 
 
 @app.route("/profile")
 @login_required
 def profile_page():
-    return render_template('profile.html', title='Profile')
+    '''All posts of a logged in user are fetching from data base'''
+    posts = Post.query.filter_by(user_id=current_user.get_id())
+    profile = User.query.filter_by(username=current_user.username).first()
+    print(profile)
+    return render_template('profile.html', title='Profile', posts=posts, profile=profile)
