+++ src/mainApp/templates/layout.html
@@ -13,6 +13,9 @@
     <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/css/bootstrap.min.css"
         integrity="sha384-ggOyR0iXCbMQv3Xipma34MD+dH/1fQ784/j6cY/iJTQUOhcWr7x9JvoRxT2MZw1T" crossorigin="anonymous">
     <link rel="stylesheet" type="text/css" href="{{ url_for('static', filename='main.css') }}">
+    {% block css %}
+
+    {% endblock %}
     <!-- change page title as needed -->
     {% if title %}
     <title>Spartan Textbooks - {{ title }}</title>
@@ -56,7 +59,7 @@
     <!-- warning message popup -->
     <main role="main" class="container">
         <div class="row">
-            <div class="col-md-8">
+            <div class="col-md-12 mt-5">
                 {% with messages =get_flashed_messages(with_categories=true) %}
                 {%if messages%}
                 {%for category, message in messages%}
